+++ src/components/userList/loadUsers.js
@@ -1,15 +1,7 @@
 import {API} from '../../api/api.js';
 import { loadTodos } from '../todoList/loadTodos.js';
 import {setCurrentUser} from '../../utils/localStorage.js';
-
-const userButtonTemplate = ({name, _id}) => {
-  return `<button class="ripple" id=${_id}>${name}</button>`;
-}
-
-const userControlTemplate = () => {
-    return `<button class="ripple user-create-button">+ 유저 생성</button>
-            <button class="ripple user-delete-button">- 유저 삭제</button>`;
-}
+import {userButtonTemplate, userControlTemplate} from './userTemplates.js';
 
 export const loadUsers = async (selectedUser = '') => {
   const users = await API.getUsers();
@@ -18,7 +10,7 @@ export const loadUsers = async (selectedUser = '') => {
   $userList.innerHTML = '';
   users.map((user) => $userList.insertAdjacentHTML('beforeend', userButtonTemplate(user)));
   $userList.insertAdjacentHTML('beforeend', userControlTemplate());
-
+  
   if(!selectedUser) {
     const firstUser = $userList.firstChild;
     firstUser.classList.add('active');
